Split BytePlus provider helpers
