+++ frontend/server/byteplus-poll.ts
@@ -7,212 +7,44 @@ import { generateAndPersistJobKeyframes } from '@/server/video-keyframes';
 import { generateAndPersistJobPreviewVideo } from '@/server/video-preview';
 import {
   BYTEPLUS_MODELARK_PROVIDER,
-  BYTEPLUS_SEEDANCE_ASPECT_RATIOS,
   PUBLIC_SEEDANCE_ENGINE_ID,
   getBytePlusArkConfig,
   getBytePlusModelArkClient,
   isBytePlusModelArkEnabled,
   scrubBytePlusError,
 } from '@/server/video-providers/byteplus-modelark';
+import {
+  expectedBytePlusTokens,
+  getBytePlusAccounting,
+  getBytePlusUnitPriceUsdPer1kTokens,
+} from './byteplus-accounting';
+import type { BytePlusPendingJob } from './byteplus-poll-types';
+import { isRecord } from './byteplus-record-utils';
+import {
+  buildNextBytePlusStorageCopyState,
+  getBytePlusStorageCopyState,
+  isBytePlusStorageCopyRetryDue,
+  resolveBytePlusStorageCopyMaxAttempts,
+  shouldRetryBytePlusStorageCopy,
+  type BytePlusStorageCopyState,
+} from './byteplus-storage-copy';
 
-type BytePlusPendingJob = {
-  job_id: string;
-  user_id: string | null;
-  engine_id: string;
-  engine_label: string;
-  provider_job_id: string;
-  status: string;
-  duration_sec: number;
-  thumb_url: string | null;
-  preview_video_url: string | null;
-  keyframe_urls: unknown;
-  aspect_ratio: string | null;
-  has_audio: boolean | null;
-  final_price_cents: number | null;
-  pricing_snapshot: unknown;
-  settings_snapshot: unknown;
-  currency: string | null;
-  payment_status: string | null;
-  updated_at: string;
-  created_at: string;
-};
+export {
+  getBytePlusAccounting,
+  getBytePlusUnitPriceUsdPer1kTokens,
+} from './byteplus-accounting';
+export type { BytePlusStorageCopyState } from './byteplus-storage-copy';
+export {
+  buildNextBytePlusStorageCopyState,
+  getBytePlusStorageCopyState,
+  isBytePlusStorageCopyRetryDue,
+  resolveBytePlusStorageCopyMaxAttempts,
+  shouldRetryBytePlusStorageCopy,
+} from './byteplus-storage-copy';
 
 const POLL_INITIAL_DELAY_MS = 5_000;
 const POLL_MAX_DURATION_MS = 35 * 60_000;
-const BYTEPLUS_OUTPUT_COPY_WINDOW_MS = 3 * 60 * 60_000;
-const DEFAULT_PROVIDER_VIDEO_COPY_MAX_ATTEMPTS = 6;
-const BYTEPLUS_STORAGE_COPY_RETRY_DELAYS_MS = [2, 5, 15, 45, 90].map((minutes) => minutes * 60_000);
-const BYTEPLUS_FAST_UNIT_PRICE_USD_PER_1K_TOKENS = 0.0056;
-const BYTEPLUS_STANDARD_UNIT_PRICE_USD_PER_1K_TOKENS = 0.007;
-// Live V1a smoke test: 720p / 16:9 / 5.041667s returned 108,900 tokens, $0.60984 provider cost.
 const ACTIVE_JOB_STATUSES = ['pending', 'queued', 'running', 'processing', 'in_progress'];
-
-const BYTEPLUS_TOKEN_DIMENSIONS: Record<string, Record<string, { width: number; height: number }>> = {
-  '480p': {
-    '21:9': { width: 1120, height: 480 },
-    '16:9': { width: 854, height: 480 },
-    '4:3': { width: 640, height: 480 },
-    '1:1': { width: 480, height: 480 },
-    '3:4': { width: 480, height: 640 },
-    '9:16': { width: 480, height: 854 },
-  },
-  '720p': {
-    '21:9': { width: 1680, height: 720 },
-    '16:9': { width: 1280, height: 720 },
-    '4:3': { width: 960, height: 720 },
-    '1:1': { width: 720, height: 720 },
-    '3:4': { width: 720, height: 960 },
-    '9:16': { width: 720, height: 1280 },
-  },
-  '1080p': {
-    '21:9': { width: 2520, height: 1080 },
-    '16:9': { width: 1920, height: 1080 },
-    '4:3': { width: 1440, height: 1080 },
-    '1:1': { width: 1080, height: 1080 },
-    '3:4': { width: 1080, height: 1440 },
-    '9:16': { width: 1080, height: 1920 },
-  },
-};
-
-function expectedBytePlusTokens(job: Pick<BytePlusPendingJob, 'duration_sec' | 'settings_snapshot'>): number {
-  const settings = isRecord(job.settings_snapshot) ? job.settings_snapshot : {};
-  const core = isRecord(settings.core) ? settings.core : {};
-  const resolution = typeof core.resolution === 'string' ? core.resolution : '720p';
-  const aspectRatio =
-    typeof core.aspectRatio === 'string' && BYTEPLUS_SEEDANCE_ASPECT_RATIOS.includes(core.aspectRatio as never)
-      ? core.aspectRatio
-      : '16:9';
-  const dimensions = BYTEPLUS_TOKEN_DIMENSIONS[resolution]?.[aspectRatio] ?? BYTEPLUS_TOKEN_DIMENSIONS['720p']['16:9'];
-  return (dimensions.width * dimensions.height * Math.max(1, Math.round(job.duration_sec)) * 24) / 1024;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === 'object' && !Array.isArray(value);
-}
-
-export type BytePlusStorageCopyState = {
-  attempts: number;
-  firstFailedAt: string | null;
-  lastFailedAt: string | null;
-  lastProviderStatus?: string | null;
-  lastReason?: string | null;
-  nextRetryAt?: string | null;
-};
-
-export function resolveBytePlusStorageCopyMaxAttempts(raw = process.env.PROVIDER_VIDEO_COPY_MAX_ATTEMPTS ?? process.env.BYTEPLUS_STORAGE_COPY_MAX_ATTEMPTS): number {
-  const parsed = Number.parseInt(String(raw ?? ''), 10);
-  return Number.isFinite(parsed) && parsed > 0 ? Math.min(parsed, 50) : DEFAULT_PROVIDER_VIDEO_COPY_MAX_ATTEMPTS;
-}
-
-export function getBytePlusStorageCopyState(settingsSnapshot: unknown): BytePlusStorageCopyState {
-  const settings = isRecord(settingsSnapshot) ? settingsSnapshot : {};
-  const raw = isRecord(settings.byteplusStorageCopy) ? settings.byteplusStorageCopy : {};
-  const attempts = Number(raw.attempts);
-  return {
-    attempts: Number.isFinite(attempts) && attempts > 0 ? Math.floor(attempts) : 0,
-    firstFailedAt: typeof raw.firstFailedAt === 'string' && raw.firstFailedAt.length ? raw.firstFailedAt : null,
-    lastFailedAt: typeof raw.lastFailedAt === 'string' && raw.lastFailedAt.length ? raw.lastFailedAt : null,
-    lastProviderStatus: typeof raw.lastProviderStatus === 'string' ? raw.lastProviderStatus : null,
-    lastReason: typeof raw.lastReason === 'string' ? raw.lastReason : null,
-    nextRetryAt: typeof raw.nextRetryAt === 'string' && raw.nextRetryAt.length ? raw.nextRetryAt : null,
-  };
-}
-
-export function buildNextBytePlusStorageCopyState(
-  settingsSnapshot: unknown,
-  params: { nowIso?: string; providerStatus?: string | null; reason?: string | null; maxAttempts?: number } = {}
-): BytePlusStorageCopyState {
-  const previous = getBytePlusStorageCopyState(settingsSnapshot);
-  const nowIso = params.nowIso ?? new Date().toISOString();
-  const attempts = previous.attempts + 1;
-  const maxAttempts = params.maxAttempts ?? resolveBytePlusStorageCopyMaxAttempts();
-  const retryDelayMs = attempts < maxAttempts ? BYTEPLUS_STORAGE_COPY_RETRY_DELAYS_MS[attempts - 1] : null;
-  const nowMs = Date.parse(nowIso);
-  return {
-    attempts,
-    firstFailedAt: previous.firstFailedAt ?? nowIso,
-    lastFailedAt: nowIso,
-    lastProviderStatus: params.providerStatus ?? previous.lastProviderStatus ?? null,
-    lastReason: params.reason ?? previous.lastReason ?? null,
-    nextRetryAt:
-      retryDelayMs !== null && Number.isFinite(nowMs) ? new Date(nowMs + retryDelayMs).toISOString() : null,
-  };
-}
-
-export function isBytePlusStorageCopyRetryDue(
-  state: Pick<BytePlusStorageCopyState, 'nextRetryAt'>,
-  nowMs = Date.now()
-): boolean {
-  if (!state.nextRetryAt) return true;
-  const nextRetryAtMs = Date.parse(state.nextRetryAt);
-  if (!Number.isFinite(nextRetryAtMs)) return true;
-  return nowMs >= nextRetryAtMs;
-}
-
-export function shouldRetryBytePlusStorageCopy(params: {
-  state: Pick<BytePlusStorageCopyState, 'attempts'>;
-  createdAt: string;
-  nowMs?: number;
-  maxAttempts?: number;
-  copyWindowMs?: number;
-}): boolean {
-  const maxAttempts = params.maxAttempts ?? resolveBytePlusStorageCopyMaxAttempts();
-  if (params.state.attempts >= maxAttempts) return false;
-  const createdAtMs = Date.parse(params.createdAt);
-  if (!Number.isFinite(createdAtMs)) return true;
-  const copyWindowMs = params.copyWindowMs ?? BYTEPLUS_OUTPUT_COPY_WINDOW_MS;
-  return (params.nowMs ?? Date.now()) - createdAtMs < copyWindowMs;
-}
-
-export function getBytePlusAccounting(job: Pick<BytePlusPendingJob, 'settings_snapshot' | 'has_audio'>) {
-  const settings = isRecord(job.settings_snapshot) ? job.settings_snapshot : {};
-  const refs = isRecord(settings.refs) ? settings.refs : {};
-  const mode =
-    settings.inputMode === 'extend'
-      ? 'extend'
-      : settings.inputMode === 'v2v'
-        ? 'v2v'
-        : settings.inputMode === 'ref2v'
-          ? 'ref2v'
-          : settings.inputMode === 'i2v'
-            ? 'i2v'
-            : 't2v';
-  const hasStartImage = mode === 'i2v' && typeof refs.imageUrl === 'string' && refs.imageUrl.trim().length > 0;
-  const hasEndImage = mode === 'i2v' && typeof refs.endImageUrl === 'string' && refs.endImageUrl.trim().length > 0;
-  const hasReferenceImages = Array.isArray(refs.referenceImages) && refs.referenceImages.length > 0;
-  const hasReferenceVideos = Array.isArray(refs.videoUrls) && refs.videoUrls.length > 0;
-  const hasReferenceAudio = typeof refs.audioUrl === 'string' || (Array.isArray(refs.audioUrls) && refs.audioUrls.length > 0);
-  const inputType =
-    mode === 'extend'
-      ? 'video_extension'
-      : mode === 'v2v'
-        ? 'video_edit'
-        : mode === 'ref2v'
-      ? 'reference_generation'
-      : hasEndImage
-        ? 'first_last_frame'
-        : hasStartImage
-          ? 'image_input'
-          : 'text_input';
-
-  return {
-    mode,
-    inputType,
-    hasStartImage,
-    hasEndImage,
-    hasReferenceImages,
-    hasReferenceVideos,
-    hasReferenceAudio,
-    generateAudio: job.has_audio === true,
-    byteplusBillingInputType: hasReferenceVideos ? 'video_input' : 'no_video_input',
-  };
-}
-
-export function getBytePlusUnitPriceUsdPer1kTokens(engineId: string | null | undefined): number {
-  return engineId === PUBLIC_SEEDANCE_ENGINE_ID
-    ? BYTEPLUS_STANDARD_UNIT_PRICE_USD_PER_1K_TOKENS
-    : BYTEPLUS_FAST_UNIT_PRICE_USD_PER_1K_TOKENS;
-}
 
 async function recordPollEvent(
   job: Pick<BytePlusPendingJob, 'job_id' | 'provider_job_id' | 'engine_id'>,
